Move View namespace to Views.
