+++ MinimalMVVM/App.xaml
@@ -1,7 +1,7 @@
 ﻿<Application x:Class="MinimalMVVM.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             StartupUri="View\ConvertWindow.xaml">
+             StartupUri="Views/ConvertWindow.xaml">
     <Application.Resources>
          
     </Application.Resources>
