+++ MinimalMVVM/Models/TextConverter.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 
 namespace MinimalMVVM.Models
 {
