+++ MinimalMVVM/Views/ConvertWindow.xaml
@@ -1,8 +1,8 @@
-<Window x:Class="MinimalMVVM.View.ConvertWindow"
+<Window x:Class="MinimalMVVM.Views.ConvertWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:ViewModel="clr-namespace:MinimalMVVM.ViewModel"
-        xmlns:View="clr-namespace:MinimalMVVM.View"
+        xmlns:Views="clr-namespace:MinimalMVVM.Views"
         Title="Converter"
         MinWidth="300"
         ResizeMode="NoResize"
@@ -12,6 +12,6 @@
         <ViewModel:Presenter/>
     </Window.DataContext>
     
-    <View:ConverterControl/>
+    <Views:ConverterControl/>
     
 </Window>
